Use a nested class for the Kotlin tests & minimize diff
