+++ src/test/java/org/openrewrite/java/spring/http/SimplifyWebTestClientCallsTest.java
@@ -16,16 +16,19 @@
 package org.openrewrite.java.spring.http;
 
 import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.openrewrite.DocumentExample;
 import org.openrewrite.InMemoryExecutionContext;
 import org.openrewrite.java.JavaParser;
+import org.openrewrite.kotlin.KotlinParser;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 
 import static org.openrewrite.java.Assertions.java;
+import static org.openrewrite.kotlin.Assertions.kotlin;
 
 class SimplifyWebTestClientCallsTest implements RewriteTest {
 
@@ -34,6 +37,8 @@ class SimplifyWebTestClientCallsTest implements RewriteTest {
         spec
           .recipe(new SimplifyWebTestClientCalls())
           .parser(JavaParser.fromJavaVersion()
+            .classpathFromResources(new InMemoryExecutionContext(), "spring-web-6", "spring-test-6"))
+          .parser(KotlinParser.builder()
             .classpathFromResources(new InMemoryExecutionContext(), "spring-web-6", "spring-test-6"));
     }
 
@@ -92,6 +97,66 @@ class SimplifyWebTestClientCallsTest implements RewriteTest {
               """.formatted(method)
           )
         );
+    }
+
+    @Nested
+    class KotlinTest {
+        @DocumentExample
+        @ParameterizedTest
+        @CsvSource({
+          "200,isOk()",
+          "201,isCreated()",
+          "202,isAccepted()",
+          "204,isNoContent()",
+          "302,isFound()",
+          "303,isSeeOther()",
+          "304,isNotModified()",
+          "307,isTemporaryRedirect()",
+          "308,isPermanentRedirect()",
+          "400,isBadRequest()",
+          "401,isUnauthorized()",
+          "403,isForbidden()",
+          "404,isNotFound()"
+        })
+        void replacesAllIntStatusCodes(String httpStatus, String method) {
+            rewriteRun(
+              //language=kotlin
+              kotlin(
+                """
+                  import org.springframework.test.web.reactive.server.WebTestClient
+
+                  class Test {
+                      val webClient: WebTestClient = WebTestClient.bindToServer().build()
+                      fun someMethod() {
+                        webClient
+                            .post()
+                            .uri("/some/url")
+                            .bodyValue("someValue")
+                            .exchange()
+                            .expectStatus()
+                            .isEqualTo(%s)
+                      }
+                  }
+                  """.formatted(httpStatus),
+                """
+                  import org.springframework.test.web.reactive.server.WebTestClient
+
+                  class Test {
+                      val webClient: WebTestClient = WebTestClient.bindToServer().build()
+                      fun someMethod() {
+                        webClient
+                            .post()
+                            .uri("/some/url")
+                            .bodyValue("someValue")
+                            .exchange()
+                            .expectStatus()
+                            .%s
+                      }
+                  }
+                  """.formatted(method)
+              )
+            );
+        }
     }
 
     @Test
